# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Clashtools Backend API
  description: API Dokumentation für das Clashtools Backend
  contact:
    name: Thomas Standegger
    url: https://htl-kaindorf.at
    email: statha21@htl-kaindorf.at
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /api/base-entities:
    get:
      tags:
        - Base Entity
      summary: Get all base entities
      description: Retrieves a list of all base entities
      operationId: getAllBaseEntities
      responses:
        '200':
          description: Successfully retrieved the list of base entities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BaseEntityDTO'
    post:
      tags:
        - Base Entity
      summary: Create a new base entity
      description: Creates a new base entity
      operationId: createBaseEntity
      requestBody:
        description: Base entity data to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseEntityRequestDTO'
        required: true
      responses:
        '200':
          description: Successfully created the base entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseEntityDTO'
        '400':
          description: Invalid input data
  /api/base-entities/{id}:
    get:
      tags:
        - Base Entity
      summary: Get base entity by ID
      description: Retrieves a base entity by its ID
      operationId: getBaseEntityById
      parameters:
        - name: id
          in: path
          description: ID of the base entity to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the base entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseEntityDTO'
        '404':
          description: Base entity not found
    put:
      tags:
        - Base Entity
      summary: Update a base entity
      description: Updates an existing base entity by its ID
      operationId: updateBaseEntity
      parameters:
        - name: id
          in: path
          description: ID of the base entity to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated base entity data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseEntityRequestDTO'
        required: true
      responses:
        '200':
          description: Successfully updated the base entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseEntityDTO'
        '400':
          description: Invalid input data
        '404':
          description: Base entity not found
    delete:
      tags:
        - Base Entity
      summary: Delete a base entity
      description: Deletes a base entity by its ID
      operationId: deleteBaseEntity
      parameters:
        - name: id
          in: path
          description: ID of the base entity to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successfully deleted the base entity
        '404':
          description: Base entity not found
...
components:
  schemas:
...
    BaseEntityDTO:
      type: object
      description: Data Transfer Object for Base Entity information
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the base entity
          example: 1
        name:
          type: string
          description: Name of the base entity
          example: Town Hall
        level:
          type: integer
          description: Level of the base entity
          example: 10
...
    BaseEntityRequestDTO:
      type: object
      description: Request DTO for creating or updating a Base Entity
      properties:
        name:
          type: string
          description: Name of the base entity
          example: Town Hall
        level:
          type: integer
          description: Level of the base entity
          example: 10
      required:
        - name
        - level
...